+++ packages/crdt_socket_sync/client_example/lib/generated.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: constant_identifier_names, non_constant_identifier_names
 
-String crdt_socket_sync_version = '0.6.1';
-String crdt_lf_version = '4.0.0';
-String crdt_lf_flutter_version = '0.3.0';
+String crdt_socket_sync_version = '0.6.0';
+String crdt_lf_version = '3.4.2+1';
+String crdt_lf_flutter_version = '0.2.1';